servers/job/requirements.txt
streamlit
openai 
pymupdf 
python-dotenv
apify-client